+++ setup.py
@@ -1,0 +1,32 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+# Copyright (c) 2024 SoftBank Corp.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Setup script for the resource management server."""
+
+
+from setuptools import find_packages
+from setuptools import setup
+
+setup(
+    name="resource_management_server",
+    version="0.0.0",
+    packages=find_packages(),
+    include_package_data=True,
+    install_requires=[
+        "Flask",
+        "pysqlite3"
+    ],
+)
